requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-openai
mysql-connector-python
